Разбивает js на модули
