+++ js/data.js
@@ -1,0 +1,28 @@
+const getData = () => {
+  const PHOTO_DESCRIPTIONS = [
+    'Сложно сказать, почему логотип крупнейшей компании по производству мыльных пузырей развеял последние сомнения',
+    'Случайный текст похож на звон колоколов',
+    'Британские «ученые» заявили, что небо темнеет',
+    'Как бы то ни было, современная методология разработки оказалась чрезвычайно полезной'
+  ];
+
+  const COMMENT_MESSAGES = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+  ];
+
+  const COMMENT_NAMES = [
+    'Роман',
+    'Дарья',
+    'Виктория',
+    'Иван'
+  ];
+
+  return {PHOTO_DESCRIPTIONS, COMMENT_MESSAGES, COMMENT_NAMES};
+};
+
+export {getData};
